Combine from_vc_jwt tests back into one test suite and test module
