--- tests/unit_test_cases/verifiable_credential_1_1_data_model_validation.json
@@ -1,11 +1,0 @@
-[
-  "test_validate_dm_empty_id",
-  "test_validate_dm_empty_context",
-  "test_validate_dm_context_base_mismatch",
-  "test_validate_dm_empty_type",
-  "test_validate_dm_type_base_mismatch",
-  "test_validate_dm_empty_issuer",
-  "test_validate_dm_empty_subject",
-  "test_validate_dm_issuance_in_future",
-  "test_validate_dm_credential_expired"
-]
